Make some gedbrowserng tests be integration tests
