+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/controller/test/ChildrenControllerIT.java
@@ -34,7 +34,7 @@ import lombok.extern.slf4j.Slf4j;
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
 @AutoConfigureRestTestClient
-class ChildrenControllerTest {
+class ChildrenControllerIT {
     /**
      * RestTestClient injected by Spring's test support.
      */
