+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/controller/test/ParentsControllerIT.java
@@ -31,7 +31,7 @@ import lombok.extern.slf4j.Slf4j;
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
 @AutoConfigureRestTestClient
-class ParentsControllerTest {
+class ParentsControllerIT {
 
     /**
      * RestTestClient injected by Spring's test support.
