+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/controller/test/SaveControllerIT.java
@@ -22,7 +22,7 @@ import org.springframework.test.web.servlet.client.RestTestClient;
     TestConfiguration.class }, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = { "management.port=0" })
 @AutoConfigureRestTestClient
-public class SaveControllerTest {
+public class SaveControllerIT {
     /**
      * RestTestClient injected by Spring's test support.
      */
