+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/controller/test/SpousesControllerIT.java
@@ -32,7 +32,7 @@ import lombok.extern.slf4j.Slf4j;
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
 @AutoConfigureRestTestClient
-class SpousesControllerTest {
+class SpousesControllerIT {
     /**
      * RestTestClient injected by Spring's test support.
      */
