+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/controller/test/SubmitterControllerIT.java
@@ -32,7 +32,7 @@ import org.springframework.web.client.RestClientException;
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = { "management.port=0" })
 @AutoConfigureRestTestClient
-public class SubmitterControllerTest {
+public class SubmitterControllerIT {
     /**
      * RestTestClient injected by Spring's test support.
      */
