+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/crud/test/ChildCrudIT.java
@@ -29,7 +29,7 @@ import lombok.extern.slf4j.Slf4j;
 @TestPropertySource(properties = {"management.port=0"})
 @SuppressWarnings("PMD.JUnitTestsShouldIncludeAssert")
 @Slf4j
-class ChildCrudTest {
+class ChildCrudIT {
 
     /** */
     @Autowired
