+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/crud/test/FamilyCrudIT.java
@@ -31,7 +31,7 @@ import lombok.extern.slf4j.Slf4j;
     TestConfiguration.class }, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
-class FamilyCrudTest {
+class FamilyCrudIT {
 
     /** */
     @Autowired
