+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/crud/test/HeadCrudIT.java
@@ -26,7 +26,7 @@ import lombok.extern.slf4j.Slf4j;
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = {"management.port=0"})
 @Slf4j
-public class HeadCrudTest {
+public class HeadCrudIT {
 
     /** */
     @Autowired
