+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/crud/test/NoteCrudIT.java
@@ -33,7 +33,7 @@ import lombok.extern.slf4j.Slf4j;
     TestConfiguration.class }, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
-public class NoteCrudTest {
+public class NoteCrudIT {
 
     /** */
     @Autowired
