+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/crud/test/PersonCrudIT.java
@@ -36,7 +36,7 @@ import lombok.extern.slf4j.Slf4j;
     TestConfiguration.class }, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
-public class PersonCrudTest {
+public class PersonCrudIT {
 
     /** */
     @Autowired
