+++ gedbrowserng/src/test/java/org/schoellerfamily/gedbrowser/api/crud/test/SpouseCrudIT.java
@@ -32,7 +32,7 @@ import lombok.extern.slf4j.Slf4j;
     TestConfiguration.class }, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @TestPropertySource(properties = { "management.port=0" })
 @Slf4j
-public final class SpouseCrudTest {
+public final class SpouseCrudIT {
 
     /** */
     @Autowired
